+++ src/lexer.mli
@@ -4,11 +4,11 @@ val escape_string : string -> string
 
 type tokenizer_state
 
-val tokenizer_state_of_string : ?fname:string -> string -> tokenizer_state
+val tokenizer_state_of_string : ?compat:[ `V1 ] -> ?fname:string -> string -> tokenizer_state
 
-val tokenizer_state_of_channel : ?fname:string -> in_channel -> tokenizer_state
+val tokenizer_state_of_channel : ?compat:[ `V1 ] -> ?fname:string -> in_channel -> tokenizer_state
 
-val tokenizer_state_of_fun : ?fname:string -> (bytes -> offset:int -> len:int -> int) -> tokenizer_state
+val tokenizer_state_of_fun : ?compat:[ `V1 ] -> ?fname:string -> (bytes -> offset:int -> len:int -> int) -> tokenizer_state
 
 val main_tokenizer : tokenizer_state -> unit -> Parser.token * Lexing.position * Lexing.position