+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## UNRELEASED
 
+- Added lexer-level compatibility for old KDLv1 documents.
 - Added `Kdl.{to_string_compact,node_to_string,node_to_string_compact}`.
 - Simplified the semantics of `Kdl.Num.equal`.
 - Fixed a token's `pos_bol` being zero.